+++ src/partitioning/qbvh/mod.rs
@@ -4,9 +4,10 @@ pub use self::build::{
 };
 
 pub use self::qbvh::{
-    GenericQbvh, IndexedData, NodeIndex, Qbvh, QbvhNode, QbvhProxy, SimdNodeIndex,
+    GenericQbvh, IndexedData, NodeIndex, Qbvh, QbvhNode, QbvhNodeFlags, QbvhProxy, SimdNodeIndex,
 };
 pub use self::storage::QbvhStorage;
+pub use self::update::QbvhUpdateWorkspace;
 
 mod qbvh;
 mod storage;